### 2020-12-18 #2922 @kthyng
@diehlpk would this be up your alley (on the math and solver side)? Are you able to edit this?
