### 2020-12-28 0975a50
Update code and manual
