### 2021-01-02 2df7653
Delete all MATLAB stuff
