### 2021-01-02 8d9c22e
Interpolation in 1d-omega space
