+++ modules/interpolateFreq.py
@@ -15,7 +15,7 @@ from scipy import interpolate
 from numpy import loadtxt
 import numpy as np
 
-def interpolateFreq(a,b,omega,Nev,filePath,npts,isSolve):
+def interpolateCoeffsFreq(a,b,omega,Nev,filePath,npts,isSolve):
     omegaNew=np.linspace(a,b,npts+1)
     H=np.zeros((Nev**2,len(omega)),dtype=complex)
     F=np.zeros((Nev,len(omega)),dtype=complex)
@@ -50,7 +50,7 @@ def interpolateFreq(a,b,omega,Nev,filePath,npts,isSolve):
     return omegaNew
 
 
-def interpolateFreqComplex(omega,omegaNew,Nev,filePath):
+def interpolateCoeffsFreqComplex(omega,omegaNew,Nev,filePath):
     omegaflat=omega.flatten(order='F')
     omeganewflat=omegaNew.flatten(order='F')
     H=np.zeros((Nev**2,len(omegaflat)),dtype=complex)
@@ -66,13 +66,13 @@ def interpolateFreqComplex(omega,omegaNew,Nev,filePath):
     FNew=np.zeros((Nev,len(omegaNew)),dtype=complex)
     for m in range(0,Nev**2-1):
         ReH=np.reshape(H[m,:],(np.shape(omega)[0],np.shape(omega)[1]))
-        f=interp2d(omegaflat.real, omegaflat.imag, ReH)
+        f=interpolate.interp2d(omegaflat.real, omegaflat.imag, ReH)
         HH=f(omeganewflat.real,omeganewflat.imag)
         HNew[m,:]=HH
 
         if(m<Nev):
             ReF=np.reshape(F[m,:],(np.shape(omega)[0],np.shape(omega)[1]))
-            f=interp2d(omegaflat.real,omegaflat.imag,ReF)
+            f=interpolate.interp2d(omegaflat.real,omegaflat.imag,ReF,kind='cubic')
             FF=f(omeganewflat.real,omeganewflat.imag)
             FNew[m,:]=FF
     np.savetxt(filePath+"Interpolated_H/ReH.dat",HNew.real,delimiter="\t",newline="\n")
@@ -86,5 +86,52 @@ def interpolateFreqComplex(omega,omegaNew,Nev,filePath):
         lambdaj[p,:]=np.linalg.solve(Hmat,Fmat)
     np.savetxt(filePath+"Interpolated_L/lambdaRe.dat",lambdaj.real,delimiter="\t",newline="\n")
     np.savetxt(filePath+"Interpolated_L/lambdaIm.dat",lambdaj.imag,delimiter="\t",newline="\n")
+    return 0
+
+def interpolateRefCoeff(omega,omegaNew,Nev,filePath,TorC):
+    rd=np.zeros((1,len(omega)),dtype=complex)
+    rr=np.zeros((len(omega),Nev),dtype=complex)
+    for m in range(0,len(omega)):
+        rcDiff=np.loadtxt(filePath+"RefCoeff_Dif/ref"+TorC+str(m+1)+".dat",delimiter="\t")
+        rcRad=np.loadtxt(filePath+"RefCoeff_Rad/ref"+TorC+str(m+1)+".dat",delimiter="\t")
+        rd[0,m]=rcDiff[0]+1j*rcDiff[1]
+        rr[m,:]=rcRad[:,0]+1j*rcRad[:,1]
+    f=interpolate.interp1d(omega,rd,kind='cubic')
+    rdNew=f(omegaNew)
+    rcNew=np.zeros((len(omegaNew),Nev),dtype=complex)
[…]
