+++ ComplexRefIce.py
@@ -25,7 +25,7 @@ for m in range(0,npts):
     for n in range(0,npts):
         cmd="/usr/local/bin/mpirun -np 2 /usr/local/ff++/mpich3/bin/FreeFem++-mpi -v 0 iceberg.edp -N1 20 -N2 80 -Tr "+str(T[m,n].real)+" -Ti "+str(T[m,n].imag)+" -L 1000 -H 5000 -h 200 -nev 8 -iter "+str(count)+" > /dev/null";
         #print(cmd)
-        #a=system(cmd)
+        a=system(cmd)
         #print("Done running",count)
         count=count+1
 for m in range(0,npts**2):
@@ -43,4 +43,41 @@ omeganew=Xq+1j*Yq;
 Tnew=2*pi/omeganew
 
 interpolateCoeffsFreqComplex(a,b,c,d,npts,8,SolutionDir+"2_ModesMatrix/",nptsNew)
+LAMRe=np.loadtxt(SolutionDir+"2_ModesMatrix/Interpolated_L/lambdaRe.dat")
+LAMIm=np.loadtxt(SolutionDir+"2_ModesMatrix/Interpolated_L/lambdaIm.dat")
+LAM=LAMRe+1j*LAMIm
+
 interpolateRefCoeffComplex(a,b,c,d,npts,8,SolutionDir+"2_RefCoeff/","C",nptsNew)
+RefCDifRe=np.loadtxt(SolutionDir+"2_RefCoeff/Interpolated_R/refCDifRe.dat");
+RefCDifIm=np.loadtxt(SolutionDir+"2_RefCoeff/Interpolated_R/refCDifIm.dat");
+RefCRadRe=np.loadtxt(SolutionDir+"2_RefCoeff/Interpolated_R/refCRadRe.dat");
+RefCRadIm=np.loadtxt(SolutionDir+"2_RefCoeff/Interpolated_R/refCRadIm.dat");
+RCDif=RefCDifRe+1j*RefCDifIm
+RCRad=RefCRadRe+1j*RefCRadIm
+RC=np.zeros(len(RCDif),dtype=complex)
+for indx in range(0,len(RC)):
+    L=LAM[indx,:]
+    RC[indx]=RCDif[indx]+np.dot(RCRad[indx,:],L)
+
+RCNew=np.reshape(RC,np.shape(omeganew))
+VAL=cplot.get_srgb1(RCNew)
+plt.subplot(1,2,1)
+plt.imshow(VAL)
+
+interpolateRefCoeffComplex(a,b,c,d,npts,8,SolutionDir+"2_RefCoeff/","T",nptsNew)
+RefTDifRe=np.loadtxt(SolutionDir+"2_RefCoeff/Interpolated_R/refTDifRe.dat");
+RefTDifIm=np.loadtxt(SolutionDir+"2_RefCoeff/Interpolated_R/refTDifIm.dat");
+RefTRadRe=np.loadtxt(SolutionDir+"2_RefCoeff/Interpolated_R/refTRadRe.dat");
+RefTRadIm=np.loadtxt(SolutionDir+"2_RefCoeff/Interpolated_R/refTRadIm.dat");
+RTDif=RefTDifRe+1j*RefTDifIm
+RTRad=RefTRadRe+1j*RefTRadIm
+RT=np.zeros(len(RTDif),dtype=complex)
+for indx in range(0,len(RC)):
+    L=LAM[indx,:]
+    RT[indx]=RTDif[indx]+np.dot(RTRad[indx,:],L)
+
+RTNew=np.reshape(RT,np.shape(omeganew))
+VAL=cplot.get_srgb1(RTNew)
+plt.subplot(1,2,2)
+plt.imshow(VAL)
+plt.show()
