### 2021-01-03 33342db
Add new functions to assemble the reflection coefficients and the solution
