+++ RealRefIce.py
@@ -1,5 +1,5 @@
 import numpy as np
-from modules.interpolateFreq import interpolateCoeffsFreq, interpolateRefCoeff
+from modules.interpolateFreq import *
 import matplotlib.pyplot as plt
 
 filePath="2_ICEBERG/"
@@ -14,10 +14,7 @@ plt.plot(omega,abs(RC[:,0]+1j*RC[:,1]),'r+')
 
 npts=299
 omeganew=interpolateCoeffsFreq(2*pi/80,2*pi/15,omega,8,filePath+"2_ModesMatrix/",npts,1)
-
-LAMRe=np.loadtxt(filePath+"2_ModesMatrix/Interpolated_L/lambdaRe.dat")
-LAMIm=np.loadtxt(filePath+"2_ModesMatrix/Interpolated_L/lambdaIm.dat")
-LAM=LAMRe+1j*LAMIm
+LAM=buildLam(filePath)
 
 plt.subplot(2,1,1)
 for m in range(0,4):
@@ -26,33 +23,15 @@ for m in range(0,4):
 
 ## Interpolating reflection coefficients
 V=interpolateRefCoeff(omega,omeganew,8,filePath+"2_RefCoeff/","C")
-RefCDifRe=np.loadtxt(filePath+"2_RefCoeff/Interpolated_R/refCDifRe.dat");
-RefCDifIm=np.loadtxt(filePath+"2_RefCoeff/Interpolated_R/refCDifIm.dat");
-RefCRadRe=np.loadtxt(filePath+"2_RefCoeff/Interpolated_R/refCRadRe.dat");
-RefCRadIm=np.loadtxt(filePath+"2_RefCoeff/Interpolated_R/refCRadIm.dat");
-RCDif=RefCDifRe+1j*RefCDifIm
-RCRad=RefCRadRe+1j*RefCRadIm
-RC=np.zeros(len(RCDif),dtype=complex)
-
-for indx in range(0,len(RC)):
-    L=LAM[indx,:]
-    RC[indx]=RCDif[indx]+np.dot(RCRad[indx,:],L)
+RC=buildRMat(LAM,filePath,"C")
 plt.subplot(2,1,2)
 plt.plot(omeganew,np.transpose(abs(RC)),linewidth=2)
 
 ## Interpolating Transmission coefficients
 V=interpolateRefCoeff(omega,omeganew,8,filePath+"2_RefCoeff/","T")
-RefTDifRe=np.loadtxt(filePath+"2_RefCoeff/Interpolated_R/refTDifRe.dat");
-RefTDifIm=np.loadtxt(filePath+"2_RefCoeff/Interpolated_R/refTDifIm.dat");
-RefTRadRe=np.loadtxt(filePath+"2_RefCoeff/Interpolated_R/refTRadRe.dat");
-RefTRadIm=np.loadtxt(filePath+"2_RefCoeff/Interpolated_R/refTRadIm.dat");
-RTDif=RefTDifRe+1j*RefTDifIm
-RTRad=RefTRadRe+1j*RefTRadIm
-RT=np.zeros(len(RTDif),dtype=complex)
-
-for indx in range(0,len(RT)):
-    L=LAM[indx,:]
-    RT[indx]=RTDif[indx]+np.dot(RTRad[indx,:],L)
+RT=buildRMat(LAM,filePath,"T")
 plt.plot(omeganew,np.transpose(abs(RT)),linewidth=2)
+
+# Check Energy Conservation
 plt.plot(omeganew,np.transpose(abs(RT)**2+abs(RC)**2),linewidth=2)
 plt.show()
