+++ ComplexRefIce.py
@@ -1,5 +1,5 @@
 import numpy as np
-from modules.interpolateFreq import interpolateCoeffsFreqComplex, interpolateRefCoeffComplex
+from modules.interpolateFreq import *
 import cplot
 import matplotlib.pyplot as plt
 from os import system
@@ -7,7 +7,7 @@ from os import system
 SolutionDir="1_ICEBERG/"
 
 npts=10;
-pi=np.pi;
+pi=np.pi
 
 a=2*pi/80
 b=2*pi/15
@@ -25,7 +25,7 @@ for m in range(0,npts):
     for n in range(0,npts):
         cmd="/usr/local/bin/mpirun -np 2 /usr/local/ff++/mpich3/bin/FreeFem++-mpi -v 0 iceberg.edp -N1 20 -N2 80 -Tr "+str(T[m,n].real)+" -Ti "+str(T[m,n].imag)+" -L 1000 -H 5000 -h 200 -nev 8 -iter "+str(count)+" > /dev/null";
         #print(cmd)
-        a=system(cmd)
+        #a=system(cmd)
         #print("Done running",count)
         count=count+1
 for m in range(0,npts**2):
@@ -43,41 +43,21 @@ omeganew=Xq+1j*Yq;
 Tnew=2*pi/omeganew
 
 interpolateCoeffsFreqComplex(a,b,c,d,npts,8,SolutionDir+"2_ModesMatrix/",nptsNew)
-LAMRe=np.loadtxt(SolutionDir+"2_ModesMatrix/Interpolated_L/lambdaRe.dat")
-LAMIm=np.loadtxt(SolutionDir+"2_ModesMatrix/Interpolated_L/lambdaIm.dat")
-LAM=LAMRe+1j*LAMIm
+LAM=buildLam(SolutionDir)
 
 interpolateRefCoeffComplex(a,b,c,d,npts,8,SolutionDir+"2_RefCoeff/","C",nptsNew)
-RefCDifRe=np.loadtxt(SolutionDir+"2_RefCoeff/Interpolated_R/refCDifRe.dat");
-RefCDifIm=np.loadtxt(SolutionDir+"2_RefCoeff/Interpolated_R/refCDifIm.dat");
-RefCRadRe=np.loadtxt(SolutionDir+"2_RefCoeff/Interpolated_R/refCRadRe.dat");
-RefCRadIm=np.loadtxt(SolutionDir+"2_RefCoeff/Interpolated_R/refCRadIm.dat");
-RCDif=RefCDifRe+1j*RefCDifIm
-RCRad=RefCRadRe+1j*RefCRadIm
-RC=np.zeros(len(RCDif),dtype=complex)
-for indx in range(0,len(RC)):
-    L=LAM[indx,:]
-    RC[indx]=RCDif[indx]+np.dot(RCRad[indx,:],L)
-
+RC=buildRMat(LAM,SolutionDir,"C")
 RCNew=np.reshape(RC,np.shape(omeganew))
-VAL=cplot.get_srgb1(RCNew)
+RCNew=RCNew.T
+VAL=cplot.get_srgb1(RCNew,colorspace="hsl",alpha=0)
 plt.subplot(1,2,1)
 plt.imshow(VAL)
 
 interpolateRefCoeffComplex(a,b,c,d,npts,8,SolutionDir+"2_RefCoeff/","T",nptsNew)
-RefTDifRe=np.loadtxt(SolutionDir+"2_RefCoeff/Interpolated_R/refTDifRe.dat");
[…]
