+++ ComplexRefIce.py
@@ -50,14 +50,16 @@ RC=buildRMat(LAM,SolutionDir,"C")
 RCNew=np.reshape(RC,np.shape(omeganew))
 RCNew=RCNew.T
 VAL=cplot.get_srgb1(RCNew,colorspace="hsl",alpha=0)
-plt.subplot(1,2,1)
-plt.imshow(VAL)
+plt.subplot(2,1,1)
+plt.title("Reflection Coefficient")
+plt.imshow(VAL,aspect=0.4)
 
 interpolateRefCoeffComplex(a,b,c,d,npts,8,SolutionDir+"2_RefCoeff/","T",nptsNew)
 RT=buildRMat(LAM,SolutionDir,"T")
 RTNew=np.reshape(RT,np.shape(omeganew))
 RTNew=RTNew.T
 VAL=cplot.get_srgb1(RTNew,colorspace="hsl",alpha=0)
-plt.subplot(1,2,2)
-plt.imshow(VAL)
+plt.subplot(2,1,2)
+plt.title("Transmission Coefficient")
+plt.imshow(VAL,aspect=0.4)
 plt.show()
