+++ README.md
@@ -1,10 +1,15 @@
 # iceFEM++
 
 ## Dependencies
-- [**FreeFem**](https://freefem.org):
-    A high level multiphysics finite element software. Full version with the `MPI` implementation and the `ffddm` module is required. The full version can be downloaded from the official website.
-- [**Bedmap2 Toolbox**]([https://github.com/rprechelt/pybedmap2](https://github.com/rprechelt/pybedmap2)): To obtain real--life shelf/cavity data. This is available as a Python toolbox from the user *rprechelt*.
-- **Python:** This branch contains the BEDMAP2 module from [https://github.com/rprechelt/pybedmap2](https://github.com/rprechelt/pybedmap2) and the interpolation routine written using Python.
+- [**FreeFem**](https://freefem.org): A high level multiphysics finite element software. Full version with the `MPI` implementation and the `ffddm` module is required. The full version can be downloaded from the official website.
+
+- [**Bedmap2 Toolbox**]: Support yet to be provided.
+
+- **Python:** Requires Python support with `numpy`, `scipy` `matplotlib` and `cplot`, the last one for generating complex plots. Can be installed using
+```shell
+pip3 install cplot
+```
+
 - **Unix based OS:** Currently this program is written assuming a Unix based Operating system. Future support for windows will be added.
 
 ## Introduction
@@ -22,7 +27,7 @@ export FF_INCLUDEPATH="$PWD/include" #Add the include/ folder into the FreeFem++
 ./genDir.sh [YOUR_DIRECTORY NAME]
 ```
 
-The `genDir.sh` script creates the directory structure used by the package to write the relevant files. This assumes that the `MATLAB` scripts in the `modules` folder is being used. For example running
+The `genDir.sh` script creates the directory structure used by the package to write the relevant files. For example running
 
 ```shell
 ./genDir.sh 1_SIMPLE5
@@ -100,7 +105,7 @@ Model | Reflection Coefficient, `R` | abs(`R`) |
 FEM | (0.482959,0.875643) | 1.000000 |
 Thin Plate| (0.493626,0.869740) | 1.000000 |
 
-All the visualization is performed using Paraview. The code produces the following outputs for the displacement field and the velocity potential in the cavity region.
+All the visualization is performed using Paraview. The code produces the following outputs for the displacement field and the velocity potential in the cavity region. **NOTE: The Thin-plate code has been validated using the `MATLAB` module located in the other branch**.
 
 | Displacement/Potential (FEM) | Displacement (ThinPlate) |
 | --- | ---- |
@@ -126,36 +131,32 @@ T=40 s | (0.475792,-0.647576) | (0.479653,0.352414) | 1.00000 |
 
 which verifies the energy conservation result. Note that the transmission coefficient `T=0` in ice--shelf case since there is no transmission. The Q-operator boundary condition is applied on both the ends while the wave forcing function is applied only on the left hand side (inlet boundary).
 
+## Python Functions
 
-## BEDMAP2 Integration
+A list of functions written using Python has been added to perform interpolation to generate the interpolated frequency-domain solutions after obtaining the finite element solutions.
 
[…]
