### 2021-01-04 #2922 @iammix
@diehlpk Hello there, and Happy New Year! Sorry for the delayed response. Yes, I am able to review this paper, however, I will be able to start reviewing this paper at the end of next week(if time is not a problem). Please inform me if I am going to be a reviewer(In this case, are there any guidelines for reviewers). Thank you in advance.
