### 2021-01-04 #2922 @diehlpk
> Dear @diehlpk, same thing here, I'll be able to review during this month. Best regards.
> […](#)
> On 04/01/2021 13:11, Konstantinos Mixios wrote: @diehlpk <https://github.com/diehlpk> Hello there, and Happy New Year! Sorry for the delayed response. Yes, I am able to review this paper, however, I will be able to start reviewing this paper at the end of next week(if time is not a problem). Please inform me if I am going to be a reviewer(In this case, are there any guidelines for reviewers). Thank you in advance. — You are receiving this because you were mentioned. Reply to this email directly, view it on GitHub <[#2922 (comment)](https://github.com/openjournals/joss-reviews/issues/2922#issuecomment-753940872)>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AOOO3OBQVZCC7FCBX4BJNR3SYGWATANCNFSM4VBLLKJA>.

@corentin-dev sounds good. Thank you for reviewing for JOSS.
