### 2021-01-04 #2922 @diehlpk
@iammix Thanks for agreeing to review for JOSS. Please find the guidelines here

https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
