### 2021-01-04 #2922 @diehlpk
@iammix and @corentin-dev

`As a reviewer, COIs are your present or previous association with any authors of a submission: recent (past four years) collaborators in funded research or work that is published; and lifetime for the family members, business partners, and thesis student/advisor or mentor. In addition, your recent (past year) association with the same organization of a submitter is a COI, for example, being employed at the same institution.`

Can you please confirm that you do not have any COI?
