### 2021-01-04 #2939 @whedon
[REVIEW]: iceFEM: A FreeFem package for wave induced ice-shelf vibrations
**Submitting author:** @Balaje (<a href="http://orcid.org/0000-0001-6181-1876">Balaje K</a>)
**Repository:** <a href="https://github.com/Balaje/iceFem" target ="_blank">https://github.com/Balaje/iceFem</a>
**Version:** v1.0.0
