### 2021-01-06 8f854d1
Add image with axes
