### 2021-01-14 b4551ea
Add option to modify Young's Modulus
