+++ RealRefIce.py
@@ -1,39 +1,43 @@
 import numpy as np
 from modules.interpolateFreq import *
 import matplotlib.pyplot as plt
+import sys
 
-filePath="1_ICEBERG/"
+filePath=sys.argv[1];
 
-T=np.linspace(15,80,27)
+T=np.linspace(10,80,27)
 pi=np.pi
 omega=2*pi/T
 RC=np.zeros((27,4),dtype=complex)
 for m in range(1,27):
     RC[m,:]=np.loadtxt(filePath+"2_RefCoeff/rc"+str(m)+".dat")
-plt.plot(omega,abs(RC[:,0]+1j*RC[:,1]),'r+')
+plt.figure(figsize=(15,8))
+plt.plot(omega/(2*pi),abs(RC[:,0]+1j*RC[:,1]),'r+')
 
-npts=299
-omeganew=interpolateCoeffsFreq(2*pi/80,2*pi/15,omega,8,filePath+"2_ModesMatrix/",npts,1)
+npts=499
+omeganew=interpolateCoeffsFreq(2*pi/80,2*pi/10,omega,8,filePath+"2_ModesMatrix/",npts,1)
 LAM=buildLam(filePath)
 
 plt.subplot(2,1,1)
 plt.title("Coefficients vs Frequency")
-for m in range(0,7):
+for m in np.arange(0,6):
     L=LAM[:,m]
-    plt.plot(2*pi/omeganew,abs(L),linewidth=2)
+    plt.plot(omeganew/(2*pi),abs(L),linewidth=2,label="$|\lambda_"+str(m+1)+"|$")
+plt.legend()
 
 ## Interpolating reflection coefficients
 V=interpolateRefCoeff(omega,omeganew,8,filePath+"2_RefCoeff/","C")
 RC=buildRMat(LAM,filePath,"C")
 plt.subplot(2,1,2)
-plt.plot(2*pi/omeganew,np.transpose(abs(RC)),linewidth=2)
+plt.plot(omeganew/(2*pi),np.transpose(abs(RC)),linewidth=2)
 
 ## Interpolating Transmission coefficients
 V=interpolateRefCoeff(omega,omeganew,8,filePath+"2_RefCoeff/","T")
 RT=buildRMat(LAM,filePath,"T")
 plt.title("Transmission and Reflection Coefficients vs Frequency")
-plt.plot(2*pi/omeganew,np.transpose(abs(RT)),linewidth=2)
+plt.plot(omeganew/(2*pi),np.transpose(abs(RT)),linewidth=2)
 
 # Check Energy Conservation
-plt.plot(2*pi/omeganew,np.transpose(abs(RT)**2+abs(RC)**2),linewidth=2)
-plt.show()
+plt.plot(omeganew/(2*pi),np.transpose(abs(RT)**2+abs(RC)**2),linewidth=2)
+print(filePath)
+plt.savefig(sys.argv[2])
