+++ ModalCoeffs.py
@@ -3,21 +3,26 @@ from modules.interpolateFreq import *
 import matplotlib.pyplot as plt
 import sys
 
-filePath=sys.argv[1];
+plt.figure(figsize=[10,6])
+for i in range(1,len(sys.argv)):
+    filePath=sys.argv[i];
+    pi=np.pi;
+    T=np.linspace(50,1000,21)
+    omega=2*pi/T
+    npts=399
+    nev=16
+    omeganew=interpolateCoeffsFreq(2*pi/1000,2*pi/50,omega,nev,filePath+"2_ModesMatrix/",npts,1)
+    LAM=buildLam(filePath)
+    ax=plt.subplot(len(sys.argv)-1,1,i)
+    heading="Coefficients vs Frequency ("+filePath+")";
+    ax.text(.5,.9,heading,
+        horizontalalignment='center',
+        transform=ax.transAxes)
+    for m in range(0,4):
+        L=LAM[:,m]
+        plt.semilogy(omeganew,abs(L),linewidth=2,label='Mode '+str(m+1))
 
-pi=np.pi;
-omega=np.linspace(2*pi*0.0001,2*pi*0.01,10)
-T=2*pi/omega;
+    plt.xlabel('$\omega$')
+    plt.legend(loc='upper right')
 
-npts=199
-nev=64
-omeganew=interpolateCoeffsFreq(2*pi*0.0001,2*pi*0.01,omega,nev,filePath+"2_ModesMatrix/",npts,1)
-LAM=buildLam(filePath)
-
-plt.subplot(2,1,1)
-plt.title("Coefficients vs Frequency")
-for m in range(0,5):
-    L=LAM[:,m]
-    plt.plot(omeganew,abs(L),linewidth=2)
-
-#plt.show()
+plt.show()
