+++ ComplexRefIce.py
@@ -7,13 +7,13 @@ from matplotlib.ticker import FormatStrFormatter
 
 SolutionDir="2_ICEBERG/"
 
-npts=10;
+npts=11;
 pi=np.pi
 
 a=2*pi/80
-b=2*pi/15
-c=-0.06
-d=0.06
+b=2*pi/10
+c=-0.04
+d=0.04
 
 x=np.linspace(a,b,npts)
 y=np.linspace(c,d,npts)
@@ -24,10 +24,10 @@ RC=np.zeros((npts**2,4))
 count=1
 for m in range(0,npts):
     for n in range(0,npts):
-        cmd="/usr/local/bin/mpirun -np 2 /usr/local/ff++/mpich3/bin/FreeFem++-mpi -v 0 iceberg.edp -N1 20 -N2 80 -Tr "+str(T[m,n].real)+" -Ti "+str(T[m,n].imag)+" -L 1000 -H 5000 -h 200 -nev 8 -iter "+str(count)+" > /dev/null";
+        cmd="/usr/local/bin/mpirun -np 2 /usr/local/ff++/mpich3/bin/FreeFem++-mpi -v 0 iceberg.edp -N 10 -Tr "+str(T[m,n].real)+" -Ti "+str(T[m,n].imag)+" -Youngs 2e9 -L 3000 -H 1000 -h 200 -nev 8 -iter "+str(count)+" -SolDir 2_ICEBERG/ -hsize 0.04 -isSplit 1 > /dev/null"
         #print(cmd)
         #a=system(cmd)
-        #print("Done running",count)
+        print("Done running",count)
         count=count+1
 for m in range(0,npts**2):
     RC[m,:]=np.loadtxt(SolutionDir+"2_RefCoeff/rc"+str(m+1)+".dat")
@@ -36,9 +36,9 @@ RCOld=np.reshape(RCOld,(npts,npts))
 TCOld=RC[:,2]+1j*RC[:,3]
 TCOld=np.reshape(TCOld,(npts,npts))
 
-nptsNew=200;
-xq=np.linspace(2*pi/80,2*pi/15,nptsNew)
-yq=np.linspace(-0.06,0.06,nptsNew)
+nptsNew=300;
+xq=np.linspace(2*pi/80,2*pi/10,nptsNew)
+yq=np.linspace(-0.04,0.04,nptsNew)
 Xq,Yq=np.meshgrid(xq,yq)
 omeganew=Xq+1j*Yq;
 Tnew=2*pi/omeganew
@@ -48,12 +48,14 @@ LAM=buildLam(SolutionDir)
 
 interpolateRefCoeffComplex(a,b,c,d,npts,8,SolutionDir+"2_RefCoeff/","C",nptsNew)
 RC=buildRMat(LAM,SolutionDir,"C")
+
+plt.figure(figsize=[10,4])
 ax=plt.subplot(2,1,1)
-genComplexPlot(RC,omeganew)
+genComplexPlot(RC,omeganew/(2*pi))
 
 interpolateRefCoeffComplex(a,b,c,d,npts,8,SolutionDir+"2_RefCoeff/","T",nptsNew)
 RT=buildRMat(LAM,SolutionDir,"T")
 ax=plt.subplot(2,1,2)
-genComplexPlot(RT,omeganew)
+genComplexPlot(RT,omeganew/(2*pi))
 
-plt.show()
+plt.savefig("ComplexPlot.pdf",bbox_inches='tight')

[…]
