+++ ComplexRefIce.py
@@ -4,16 +4,17 @@ import cplot
 import matplotlib.pyplot as plt
 from os import system
 from matplotlib.ticker import FormatStrFormatter
+import sys
 
-SolutionDir="2_ICEBERG/"
+SolutionDir=sys.argv[1];
 
 npts=11;
 pi=np.pi
 
 a=2*pi/80
 b=2*pi/10
-c=-0.04
-d=0.04
+c=-0.08
+d=0.08
 
 x=np.linspace(a,b,npts)
 y=np.linspace(c,d,npts)
@@ -24,7 +25,7 @@ RC=np.zeros((npts**2,4))
 count=1
 for m in range(0,npts):
     for n in range(0,npts):
-        cmd="/usr/local/bin/mpirun -np 2 /usr/local/ff++/mpich3/bin/FreeFem++-mpi -v 0 iceberg.edp -N 10 -Tr "+str(T[m,n].real)+" -Ti "+str(T[m,n].imag)+" -Youngs 2e9 -L 3000 -H 1000 -h 200 -nev 8 -iter "+str(count)+" -SolDir 2_ICEBERG/ -hsize 0.04 -isSplit 1 > /dev/null"
+        cmd="/usr/local/bin/mpirun -np 2 /usr/local/ff++/mpich3/bin/FreeFem++-mpi -v 0 iceberg.edp -N 10 -Tr "+str(T[m,n].real)+" -Ti "+str(T[m,n].imag)+" -Youngs 2e8 -L 3000 -H 1000 -h 200 -nev 8 -iter "+str(count)+" -SolDir 3_ICEBERG/ -hsize 0.04 -isSplit 1 > /dev/null"
         #print(cmd)
         #a=system(cmd)
         print("Done running",count)
@@ -38,7 +39,7 @@ TCOld=np.reshape(TCOld,(npts,npts))
 
 nptsNew=300;
 xq=np.linspace(2*pi/80,2*pi/10,nptsNew)
-yq=np.linspace(-0.04,0.04,nptsNew)
+yq=np.linspace(-0.08,0.08,nptsNew)
 Xq,Yq=np.meshgrid(xq,yq)
 omeganew=Xq+1j*Yq;
 Tnew=2*pi/omeganew
