### 2021-02-10 #2939 @corentin-dev
Dear Patrick,
Sorry, I thought I had accepted the invitation, but I didn't. Can you reinvite me again please ?
I get : "This invitation has expired"
Best regards,
Corentin
PS: I used reply but this did not seem to get through.
PS2: And previously 04/02 : "Sorry, I did not give any update. As many of us, I am running behind schedule. Next week I'll be able to start the review."
