### 2021-02-10 #2939 @corentin-dev
@Balaje comments on the manuscript:
l16: on "the" finite
l20: analycity
l22: further l20, further ; consider changing one or the other
l29: consider citing ffddm (this reference maybe ? M. Bonazzoli, V. Dolean, I. G. Graham, E. A. Spence, P.-H. Tournier. Domain decomposition preconditioning for the high-frequency time-harmonic Maxwell equations with absorption. Mathematics of Computation, 2019. DOI: https://doi.org/10.1090/mcom/3447)
