+++ RealRefIce.py
@@ -16,7 +16,7 @@ nev=16
 omeganew=interpolateCoeffsFreq(2*pi/100,2*pi/10,omega,nev,filePath+"2_ModesMatrix/",npts,1)
 LAM=buildLam(filePath)
 
-plt.figure(figsize=[10,4])
+plt.figure(figsize=[5,4])
 plt.title("Modal Amplitudes")
 for m in np.arange(0,6):
     L=LAM[:,m]
