+++ paper.md
@@ -29,9 +29,9 @@ bibliography: paper.bib
 
 Vibrations of ice-shelves in response to ocean waves were first investigated by [@holdsworth1978iceberg] who proposed that resonant vibrations lead to icebergs calving from the shelf front. Since then seismometric measurements on the Ross ice-shelf, the largest Antarctic ice-shelf, confirmed the presence of this ocean wave-induced ice-shelf vibration [@bromirski2015ross; @Massom2018]. The period of vibration ranged from the long infragravity/tsunami waves to shorter, swell waves. More recently, [@brunt2011antarctic] presented the first observational evidence that a Northern Hemisphere tsunami triggered calving on the Sulzberger ice-shelf. Mathematical models based on linear wave theory have been proposed to study these ocean-wave induced ice-shelf vibrations.
 
-The numerical method is based on finite element method and the modal expansion technique. This results in a system of equations
+The numerical method is based on the finite element method and the modal expansion technique. This results in a system of equations
 $$[\mathbf{K}-\omega^2\mathbf{M}+i\omega\mathbf{B}]\{\lambda\} = \{\mathbf{f}\}$$
-whose entries are analytic functions of the incident frequency $\omega$. The dimension of the above linear system is much smaller than the finite element degrees of freedom and can be solved efficiently. Further, using the analyticity of the resulting system, a large number of frequency domain solutions can be obtained by interpolating the linear system without having to solve the much larger finite element problem on a finer frequency grid as shown in \autoref{fig:1}. Further details on the usage of the package can be found in the ``README.md`` file located in the repository page.
+whose entries are analytic functions of the incident frequency $\omega$. The dimension of the above linear system is much smaller than the finite element degrees of freedom and can be solved efficiently. Using the analyticity of the resulting system, a large number of frequency domain solutions can be obtained by interpolating the linear system without having to solve the much larger finite element problem on a finer frequency grid as shown in \autoref{fig:1}. Further details on the usage of the package can be found in the ``README.md`` file located in the repository page.
 
 ![Figure showing (Top) the value of the reflection coefficients on a coarse $\omega$-space (blue,+) and on a fine $\omega$-space (red,solid) obtained after solving the interpolated system. (Middle) The value of the reflection and transmission coefficients as a function of the incident frequency. The energy conservation result $|T|^2+|R|^2=1$ is also verified. (Bottom) Modal contribution, $|\lambda_j|$ of the various in-vacuo modes as a function of frequency. \label{fig:1}](Images/PaperImages/coeffs.png)
 
