### 2021-02-16 #2939 @Balaje
@corentin-dev 👋 Thanks for going through the manuscript. Regarding comment:
#2 'Analyticity' refers to complex analytic functions.
#4 ffddm is the module available in FreeFem (cited in the paper) to implement domain decomposition methods.
