+++ paper.md
@@ -37,6 +37,6 @@ whose entries are analytic functions of the incident frequency $\omega$. The dim
 
 
 # Statement of need
-FreeFem [@ffpp] is an open-source programming language to implement finite element methods and is based on C++. FreeFem is an ideal choice for studying ice-shelf vibrations problems due to its flexibility and ease of implementation. `iceFEM` is a FreeFem package for simulating ice-shelf vibrations, heavily inspired by the `ffddm` module available in FreeFem. Numerical Methods have been proposed to study the vibrations of these ice-shelves, predominantly based on the thickness averaged thin-plate model for the ice-shelf and the depth averaged shallow-water models for the fluid flow in the sub--shelf cavity region [@Sergienko2013; @Meylan2017]. More recently, finite element methods have been proposed to model non-uniform shelf/cavity regions [@Ilyas2018; @kalyanaraman2020coupled]. The method also involves implementing a general non-local boundary condition based on analytic expressions to handle semi-infinite regions. `iceFEM` also implements the method to solve the vibration problem for complex valued incident frequencies which are useful to study resonances in the complex plane [@kalyanaraman2020coupled]. Real-life examples using the BEDMAP2 dataset can also be imported and solved using `iceFEM`. The `iceFEM` package was validated using the thin-plate solutions obtained using the eigenfunction matching methods in [@KALYANARAMAN2019].
+FreeFem [@ffpp] is an open-source programming language to implement finite element methods and is based on C++. FreeFem is an ideal choice for studying ice-shelf vibrations problems due to its flexibility and ease of implementation. `iceFEM` is a FreeFem package for simulating ice-shelf vibrations, heavily inspired by the `ffddm` module available in FreeFem for implementing the domain decomposition methods. Numerical Methods have been proposed to study the vibrations of these ice-shelves, predominantly based on the thickness averaged thin-plate model for the ice-shelf and the depth averaged shallow-water models for the fluid flow in the sub--shelf cavity region [@Sergienko2013; @Meylan2017]. More recently, finite element methods have been proposed to model non-uniform shelf/cavity regions [@Ilyas2018; @kalyanaraman2020coupled]. The method also involves implementing a general non-local boundary condition based on analytic expressions to handle semi-infinite regions. `iceFEM` also implements the method to solve the vibration problem for complex valued incident frequencies which are useful to study resonances in the complex plane [@kalyanaraman2020coupled]. Real-life examples using the BEDMAP2 dataset can also be imported and solved using `iceFEM`. The `iceFEM` package was validated using the thin-plate solutions obtained using the eigenfunction matching methods in [@KALYANARAMAN2019].
 
 ## References
