### 2021-02-23 #2939 @iammix
@Balaje, You have done a great job. You demonstrate clearly how to implement and execute your examples in your documentation, something which is important for tools like iceFEM.
My only comment is on your manuscript.  It would help if you numbered the equations that you have.

@diehlpk I am finished with my review. I will cross out all of the boxes from my checklist when the author finishes with my comment.

Kind Regards.
