### 2021-02-23 #2939 @diehlpk
@Balaje Please add the following DOI to each of your references:

```
- 10.1038/274464a0 may be a valid DOI for title: Iceberg calving from floating glaciers by a vibrating mechanism
- 10.3189/002214311798043681 may be a valid DOI for title: Antarctic ice-shelf calving triggered by the Honshu (Japan) earthquake and tsunami, March 2011
- 10.1016/j.jfluidstructs.2020.103074 may be a valid DOI for title: A coupled fluid-elasticity model for the wave forcing of an ice-shelf
- 10.1016/j.jfluidstructs.2018.03.010 may be a valid DOI for title: Time-domain and modal response of ice shelves to wave forcing using the finite element method
- 10.3189/2013jog12j096 may be a valid DOI for title: Normal modes of a coupled ice-shelf / sub-ice-shelf cavity system
- 10.1016/j.wavemoti.2019.04.004 may be a valid DOI for title: On the shallow-water limit for modelling ocean-wave induced ice-shelf vibrations
- 10.1515/jnum-2012-0013 may be a valid DOI for title: New development in FreeFem++
- 10.1038/s41586-018-0212-1 may be a valid DOI for title: Antarctic ice shelf disintegration triggered by sea ice loss and ocean swell
```
