### 2021-02-23 #2939 @corentin-dev
The package needs some addition to the documentation :
- figure 1 explains the problem which is solved, but having the general algorithm would be nice as well in the documentation or the paper (general ideas, pointing to other papers for example) ;
- explain further how the script .edp decides which case to run, or where to save files, which file is read as input and how genDir.sh works in that matter ;
- maybe genDir.sh could have an .edp as argument and change SolutionDir while creating folders ? another solution could be that the script creates the structure of folders from the .edp ;
- what are the files outputted ? what are they describing ? (you explain for directories but not files) what are the units ? at least for the most useful ones ;
- perhaps explain further if there are difference in algorithms between the different .edp present in the package.

Not a change to be made, but a general comment about your git history : you should consider having more information in your commit messages, including what has been fixed for example (for the future of the package).

Otherwise, the package is well presented, and the paper is clear !
