### 2021-02-24 1e5af4f
Update manual and paper with DOI
