+++ paper.bib
@@ -3,7 +3,8 @@
   author={Holdsworth, G and Glynn, J},
   year={1978},
   journal={Nature},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/274464a0}
 }
 @article{bromirski2015ross,
   title={{Ross ice shelf vibrations}},
@@ -13,7 +14,8 @@
   number={18},
   pages={7589--7597},
   year={2015},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/2015GL065284}
 }
 @article{brunt2011antarctic,
   title={{Antarctic ice-shelf calving triggered by the Honshu (Japan) earthquake and tsunami, March 2011}},
@@ -23,7 +25,8 @@
   number={205},
   pages={785--788},
   year={2011},
-  publisher={International Glaciological Society}
+  publisher={International Glaciological Society},
+  doi={10.3189/002214311798043681}
 }
 @article{kalyanaraman2020coupled,
   title={A coupled fluid-elasticity model for the wave forcing of an ice-shelf},
@@ -31,6 +34,7 @@
   journal={Journal of Fluids and Structures},
   volume={97},
   year={2020},
+  doi={10.1016/j.jfluidstructs.2020.103074}
 }
 @article{Ilyas2018,
 author = {Ilyas, M. and Meylan, M. H. and Lamichhane, B. and Bennetts, L. G.},
@@ -38,7 +42,8 @@ journal = {J. Fluids Struct.},
 pages = {113--131},
 title = {{Time-domain and modal response of ice shelves to wave forcing using the finite element method}},
 volume = {80},
-year = {2018}
+year = {2018},
+doi = {10.1016/j.jfluidstructs.2018.03.010}
 }
 @article{Sergienko2013,
 author = {Sergienko, O. V.},
@@ -47,7 +52,8 @@ number = 213,
 pages = {76--80},
 title = {{Normal modes of a coupled ice-shelf / sub-ice-shelf cavity system}},
 volume = 59,
-year = 2013
+year = 2013,
+doi={10.3189/2013jog12j096}
 }
 @article{Meylan2017,
 author = {Meylan, M. H. and Bennetts, L. G. and Hosking, R. J. and Catt, E.},
@@ -56,7 +62,8 @@ number = {240},
 pages = {751--754},
 title = {{On the calculation of normal modes of a coupled ice-shelf/sub-ice-shelf cavity system}},
 volume = {63},
-year = {2017}
+year = {2017},
+doi = {10.1017/jog.2017.27}
 }
 
 @article{KALYANARAMAN2019,
@@ -65,7 +72,8 @@ journal = "Wave Motion",
 volume = 90,
 pages = "1 - 16",
 year = 2019,
-author = {Kalyanaraman, B and Bennetts, L. G. and Lamichhane, B. P. and Meylan, M. H.}
+author = {Kalyanaraman, B and Bennetts, L. G. and Lamichhane, B. P. and Meylan, M. H.},
+doi = {10.1016/j.wavemoti.2019.04.004}
 }
 
 @article{ffpp,
@@ -76,7 +84,8 @@ author = {Kalyanaraman, B and Bennetts, L. G. and Lamichhane, B. P. and Meylan, 
   NUMBER = {3-4}, PAGES = {251--265},
   ISSN = {1570-2820},
   MRCLASS = {65Y15},
-  MRNUMBER = {3043640}
+  MRNUMBER = {3043640},
+  doi = {10.1515/jnum-2012-0013}
 }
 
 @article{Massom2018,
@@ -87,5 +96,6 @@ pages = {383--389},
 publisher = {Springer US},
[…]
