### 2021-02-24 51aea8e
Paper: Remove tilde from equation
