+++ paper.md
@@ -33,7 +33,7 @@ The numerical method is based on the finite element method and the modal expansi
 \begin{equation}\label{eq:1}
 [\mathbf{K}-\omega^2\mathbf{M}+i\omega\mathbf{B}]\{\lambda\} = \{\mathbf{f}\}
 \end{equation}
-whose entries are analytic functions of the incident frequency $\omega$. The dimension of the linear system~(\autoref{eq:1}) is much smaller than the finite element degrees of freedom and can be solved efficiently. Using the analyticity of the resulting system, a large number of frequency domain solutions can be obtained by interpolating the linear system without having to solve the much larger finite element problem on a finer frequency grid as shown in \autoref{fig:1}. Further details on the usage of the package can be found in the ``README.md`` file located in the repository page.
+whose entries are analytic functions of the incident frequency $\omega$. The dimension of the linear system (\autoref{eq:1}) is much smaller than the finite element degrees of freedom and can be solved efficiently. Using the analyticity of the resulting system, a large number of frequency domain solutions can be obtained by interpolating the linear system without having to solve the much larger finite element problem on a finer frequency grid as shown in \autoref{fig:1}. Further details on the usage of the package can be found in the ``README.md`` file located in the repository page.
 
 ![Figure showing (Top) the value of the reflection coefficients on a coarse $\omega$-space (blue,+) and on a fine $\omega$-space (red,solid) obtained after solving the interpolated system. (Middle) The value of the reflection and transmission coefficients as a function of the incident frequency. The energy conservation result $|T|^2+|R|^2=1$ is also verified. (Bottom) Modal contribution, $|\lambda_j|$ of the various in-vacuo modes as a function of frequency. \label{fig:1}](Images/PaperImages/coeffs.png)
 
