### 2021-02-24 #2939 @Balaje
Hi @iammix,

Thank you for the comments and for your time on the paper. I have updated the paper with the equation number.

Regards,
Balaje

@diehlpk, I have added the DOIs to the paper.
