### 2021-02-24 #2939 @Balaje
Hi @corentin-dev,

Thanks for your comments.  I have updated the PDF manual here https://github.com/Balaje/iceFem/blob/ParIceFem/manual.pdf with your suggestions. Here are my responses:

1. I have added a small introduction to the solution method we use along with a reference to the paper. It’s in subsection 1.2 in the manual.

2.  The directory structure is mostly to organise things a bit more.  For example, the LHS (matrix H) and RHS (vector F) of the  reduced system is stored in `ROOT_DIR/2_ModesMatrix/`. The interpolated system is stored in the `ROOT_DIR/2_ModesMatrix/Interpolated_X` folder with `X` being `H,F or L`. If you use the MATLAB modules located in the modules directory, for example, the routine `interpolateFreq()` reads in the individual files from the finite element solution and writes the consolidated H, F and the solution $\lambda$ into the `Interpolated_*` directory. The user can then use this to plot the individual \lambda_j’s and so on. The same is done for the reflection coefficients as well. I have added a short description of the three main routines in the manual  in Section 3 (Quantities of Interest) of the manual. Similar thing with the `readEigen` and the `writeEigen` macros which uses the [ROOT_DIR/2_Modes] directory, the details of which you can find in the manual as well.

3. I have added a macro `setupWorkingDir(DirName)` to do this in ‘macros.idp’: https://github.com/Balaje/iceFem/blob/51aea8eb7a21ba67cfcc408d9ef81a280f979724/include/macros.idp#L102

4. For this, again refer to Section 3 in the manual, where I have added some explanation. For more details on the reflection coefficients and complex frequencies, I would point the readers to the papers

	- https://doi.org/10.1016/j.jfluidstructs.2020.103074
	- https://doi.org/10.1016/j.jfluidstructs.2018.03.010

5. The algorithm in the test examples are all the same, just the nature of the problem differs.

   - `iceshelf2d.edp` solves the 2d ice-shelf problem with the clamped condition on one end.
   - The same algorithm is used to solve motion of icebergs, which is done in `iceberg.edp`. The main difference here is that the clamped boundary condition is replaced by the free-boundary condition which could be easily generalised in FEM.
   - By changing the dimension macro (borrowed from the `ffddm` module) at the beginning of the code, i.e.,
     ```
     macro dimension 2//EOM” (or)
     macro dimension 3//EOM”
     ```
     one could easily generalise the algorithm to 3D cases as well. I borrowed some of the macros, for example `intN(Th)` from  FreeFem to handle n-dimensional integral in which the user specifies the dimension of the problem. The user then needs to update the mesh with a 3D mesh rather than a 2D mesh. This could be changed by updating the `ThIce` and `ThCavity` variable in the package with user’s mesh.

   - The BEDMAP2 implementation is the same, but the associated profiles needs to be obtained from the bedmap2 dataset (available in MATLAB only). The meshes are constructed using the `iceBEDMAP2()` macro defined in the macros.idp file.

  Thanks for your comment on the Git history, I will keep an eye out for it next time.

  Apologies for the lengthy comment. Hope things are clear.

  Regards,
  Balaje
