### 2021-03-09 #2939 @corentin-dev
> @corentin-dev I just checked and three checkboxes below functionality are unchecked. Could you please have a look?

You are right. They are now checked!
