### 2021-03-15 #2939 @diehlpk
@Balaje

- [ ] Please add the city and province/state to the affiliations.
- [ ] In Equation 1 please describe the quantities: K, M, B, f, and lambda briefly since not all people know these symbols.
- [ ] Please add a link to the README.md in the last line of the first page.
- [ ] I think it would help to give a brief definition of the coarse and fine space in the caption of Figure 1
- [ ] Can we write that FreeFem is a domain specific language based on the C++ language. I do not think it is a programming language as Python or C++.
- [ ] The related work on the science side is good, however the related work on fem packages is still missing. Can you add more details about other libraries?
- [ ]` freeFem is an ideal choice for studying ice-shelf vibrations27problems due to its flexibility and ease of implementation.` Can you elaborate here and make some claims why others FEM software has not this features, These featueres are claimed by many FEM libraries.
