### 2021-03-15 #2939 @diehlpk
@Balaje

I checked the references, and sometimes you use `Journal of Fluids and Structures` and sometimes you use `J. Fluids Struct.,`.

It is not important which you use, but please be consistent and only use one type for all references.
