+++ paper.md
@@ -16,9 +16,9 @@ authors:
   - name: Luke G. Bennetts
     affiliation: 2
 affiliations:
-  - name: University of Newcastle, Australia
+  - name: School of Mathematical and Physical Sciences, University of Newcastle, Callaghan, NSW 2308, Australia
     index: 1
-  - name: University of Adelaide, Australia
+  - name: School of Mathematical Sciences, University of Adelaide, Adelaide, SA 5005, Australia
     index: 2
 
 date: 18 December 2020
@@ -27,18 +27,211 @@ bibliography: paper.bib
 
 # Summary
 
-Vibrations of ice-shelves in response to ocean waves were first investigated by [@holdsworth1978iceberg] who proposed that resonant vibrations lead to icebergs calving from the shelf front. Since then seismometric measurements on the Ross ice-shelf, the largest Antarctic ice-shelf, confirmed the presence of this ocean wave-induced ice-shelf vibration [@bromirski2015ross; @Massom2018]. The period of vibration ranged from the long infragravity/tsunami waves to shorter, swell waves. More recently, [@brunt2011antarctic] presented the first observational evidence that a Northern Hemisphere tsunami triggered calving on the Sulzberger ice-shelf. Mathematical models based on linear wave theory have been proposed to study these ocean-wave induced ice-shelf vibrations.
+Vibrations of ice-shelves in response to ocean waves were first
+investigated by [@holdsworth1978iceberg] who proposed that resonant
+vibrations lead to icebergs calving from the shelf front. Since then
+seismometric measurements on the Ross ice-shelf, the largest Antarctic
+ice-shelf, confirmed the presence of this ocean wave-induced ice-shelf
+vibration [@bromirski2015ross; @Massom2018]. The period of vibration
+ranged from the long infragravity/tsunami waves to shorter, swell
+waves. More recently, [@brunt2011antarctic] presented the first
+observational evidence that a Northern Hemisphere tsunami triggered
+calving on the Sulzberger ice-shelf. Mathematical models based on
+linear wave theory have been proposed to study these ocean-wave
+induced ice-shelf vibrations.
 
-The numerical method is based on the finite element method and the modal expansion technique. This results in a system of equations
+![Geometry and the Governing
+equations. \label{fig:0}](Images/PaperImages/iceGeo.png)
+
+Consider the geometry and the schematic shown in \autoref{fig:0}. The
+motion of the ice-shelf (solid) is governed by the elastodynamic
+equations which are coupled with the fluid motion, governed using the
+linear potential flow theory. The fluid is modelled as a semi-infinite
+rectangular region of uniform depth, whereas the sub-shelf cavity
+region is assumed to be non-uniform. The time-domain problem is
+converted to the frequency domain by applying a transformation
+\begin{equation}
+  \Phi(x,z,t) = \text{Re}\left\{ \phi(x,z)e^{-i \omega t}
+  \right\},\quad
+  \mathbf{u}(x,z,t) = \text{Re}\left\{ \eta(x,z)e^{-i \omega t} \right\}
+\end{equation}
+at a prescribed frequency $\omega$.
+
+The semi-infinite region is truncated by constructing analytic
+expressions and deriving a non-local boundary condition of the form
+\begin{equation}
+  \partial_x \phi = \mathbf{Q}\phi + \chi \quad \text{on} \quad \Gamma_f^{(4)}.
+\end{equation}
+The computational domain is then restricted to the sub-shelf cavity
[…]
