+++ paper.bib
@@ -31,7 +31,7 @@
 @article{kalyanaraman2020coupled,
   title={A coupled fluid-elasticity model for the wave forcing of an ice-shelf},
   author={Kalyanaraman, B. and Meylan, M. H. and Bennetts, L. G. and Lamichhane, B. P.},
-  journal={Journal of Fluids and Structures},
+  journal={J. Fluids Struct.},
   volume={97},
   year={2020},
   doi={10.1016/j.jfluidstructs.2020.103074}
@@ -98,4 +98,45 @@ title = {{Antarctic ice shelf disintegration triggered by sea ice loss and ocean
 volume = {558},
 year = {2018},
 doi = {10.1038/s41586-018-0212-1}
+}
+@article{dealII92,
+  title     = {The \texttt{deal.II} Library, Version 9.2},
+  author    = {Daniel Arndt and Wolfgang Bangerth and Bruno Blais and
+               Thomas C. Clevenger and Marc Fehling and Alexander V. Grayver and
+               Timo Heister and Luca Heltai and Martin Kronbichler and Matthias Maier and
+               Peter Munch and Jean-Paul Pelteret and Reza Rastak and
+               Ignacio Thomas and Bruno Turcksin and Zhuoran Wang and David Wells},
+  journal   = {J. Numer. Math.},
+  publisher = {De Gruyter},
+  year      = {2020},
+  volume    = {28},
+  number    = {3},
+  pages     = {131-146},
+  DOI       = {10.1515/jnma-2020-0043}, 
+}
+@book{LoggMardalEtAl2012a,
+  title = {Automated Solution of Differential Equations by the Finite Element Method},
+  author = {Anders Logg and Kent-Andre Mardal and Garth N. Wells and others},
+  year = {2012},
+  publisher = {Springer},
+  doi = {10.1007/978-3-642-23099-8},
+  isbn = {978-3-642-23098-1},
+}
+@article{Badia2020,
+  doi = {10.21105/joss.02520},
+  url = {https://doi.org/10.21105/joss.02520},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {52},
+  pages = {2520},
+  author = {Santiago Badia and Francesc Verdugo},
+  title = {Gridap: An extensible Finite Element toolbox in Julia},
+  journal = {J. Open Source Softw}
+}
+@book{wegert2012visual,
+  title={Visual complex functions: an introduction with phase portraits},
+  author={Wegert, Elias},
+  year={2012},
+  publisher={Springer Science \& Business Media}
 }
