+++ paper.md
@@ -28,13 +28,13 @@ bibliography: paper.bib
 # Summary
 
 Vibrations of ice-shelves in response to ocean waves were first
-investigated by [@holdsworth1978iceberg] who proposed that resonant
+investigated by @holdsworth1978iceberg who proposed that resonant
 vibrations lead to icebergs calving from the shelf front. Since then
 seismometric measurements on the Ross ice-shelf, the largest Antarctic
 ice-shelf, confirmed the presence of this ocean wave-induced ice-shelf
 vibration [@bromirski2015ross; @Massom2018]. The period of vibration
 ranged from the long infragravity/tsunami waves to shorter, swell
-waves. More recently, [@brunt2011antarctic] presented the first
+waves. More recently, @brunt2011antarctic presented the first
 observational evidence that a Northern Hemisphere tsunami triggered
 calving on the Sulzberger ice-shelf. Mathematical models based on
 linear wave theory have been proposed to study these ocean-wave
@@ -55,10 +55,9 @@ converted to the frequency domain by applying a transformation
   \right\},\quad
   \mathbf{u}(x,z,t) = \text{Re}\left\{ \eta(x,z)e^{-i \omega t} \right\}
 \end{equation}
-at a prescribed frequency $\omega$.
-
-The semi-infinite region is truncated by constructing analytic
-expressions and deriving a non-local boundary condition of the form
+at a prescribed frequency $\omega$. The semi-infinite region is
+truncated by constructing analytic expressions and deriving a
+non-local boundary condition of the form
 \begin{equation}
   \partial_x \phi = \mathbf{Q}\phi + \chi \quad \text{on} \quad \Gamma_f^{(4)}.
 \end{equation}
@@ -67,8 +66,8 @@ region where the finite element method is used to solve the resulting
 governing equations. The finite dimensional weak formulation of
 the coupled problem is to find $(\phi_h,\mathbf{w}_h) \in V_h\times
 W_h$ such that
-\begin{eqnarray}
-\left(\nabla\phi_h,\nabla\psi\right)_{\Omega_f}
+\begin{align}
+    \left(\nabla\phi_h,\nabla\psi\right)_{\Omega_f}
     &=-i\omega\left\langle\mathbf{w}_h,\psi\right\rangle_{\Gamma_f^{(3)}} +
     \left\langle\mathbf{Q}\phi_h,\psi\right\rangle_{\Gamma_f^{(4)}} +
     \left\langle\chi,\psi\right\rangle_{\Gamma_f^{(4)}} \label{eq:1new}\\
@@ -76,19 +75,22 @@ W_h$ such that
     &=\rho_s\omega^2\left(\mathbf{w}_h,\mathbf{v}\right)_{\Omega_s} +
     \left\langle\mathbf{w}_h,\mathbf{v}\cdot\mathbf{n}\right\rangle_{\Gamma_s^{(1)}}
     -i\omega\left\langle\phi_h,\mathbf{v}\right\rangle_{\Gamma_f^{(3)}}
-\end{eqnarray}
-where $V_h$ and $W_h$ are appropriate finite element spaces.
+\end{align}
+where $V_h$ and $W_h$ are appropriate finite element spaces. For more
+details on the construction of the non-local boundary condition, refer
+to @Ilyas2018.
 
-## Numerical Method
+## Solution Method
 
-The numerical method is based on the modal expansion technique. The
-displacement and the velocity potential can be written as
+The solution method is based on the modal expansion technique
+[@Ilyas2018; @kalyanaraman2020coupled]. The displacement and the
+velocity potential can be written as
[…]
