### 2021-03-23 #2939 @Balaje
Hi @diehlpk. Thanks for the comments and sorry for the delay. For some reason, I couldn't tick the boxes... Here are my responses to your comments.

- I have updated the author list with the city names/state.
- I have added some description of the quantities and also added a brief explanation of the solution method.
- Added a link to the README at the end of the **Summary** section.
- I have added an example to describe the coarse and fine frequency spaces.
- I have modified the description of FreeFem at the start of the section.

Points 6) and 7)
- I currently am not aware of any libraries that do the sort of non-local boundary truncation which involves constructing a Dirichlet to Neumann mapping on the ocean-cavity interface after computing the wave-numbers. The mapping is obtained by constructing exact solutions using the separation of variables, where the ocean is assumed to be uniformly deep. Also, the solution can be analytically extended to the complex plane shown in Figure 3 in the paper (newly added), where the input frequencies are complex numbers. These features of the model require an automated finite element package for easy but flexible implementation.
- FreeFem is one of the best choices for this for the following reason:
  1) Good support for automated mesh generation in 2D as well as 3D.
  2) Easy to describe the weak formulation and matrix construction.
  3) Support for a variety of finite elements and linear solvers.
  4) Parallelization using MPI and integration with PETSc (I am currently working on using the latter).

  There are other excellent choices for this as well, like FEniCS, deal.II etc, but I found that the implementation is easy in FreeFem, especially for someone new to FEM. So I decided to use it to build this package. I have added some references to the other solvers in the **Statement of Need** section.

I have also made the journal names in the references uniform.

Thanks,
Balaje
